+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to **Pipecat Flows** will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.20] - 2025-08-27
+
+### Fixed
+
+- Corrected an issue where the Flow Manager state was marked as private. To
+  make this clear, added a `@property` to access the state.
+
 ## [0.0.19] - 2025-08-25
 
 ### Deprecated